+++ open-sphere-base/mantle/src/main/java/io/opensphere/mantle/util/dynenum/impl/DynamicEnumerationRegistryImpl.java
@@ -169,7 +169,7 @@ public class DynamicEnumerationRegistryImpl implements DynamicEnumerationRegistr
             while (iter.hasNext())
             {
                 iter.advance();
-                typeToIdMap.put(iter.key(), iter.value());
+                typeToIdMap.put(iter.key(), Short.valueOf(iter.value()));
             }
         }
         finally
